+++ packages/react/stories/all.stories.js
@@ -9,7 +9,9 @@ import { demoGroups } from "@carbon/charts/demo/data";
 // Loop through all demo groups
 demoGroups.forEach(demoGroup => {
 	// Create story group for each demo group
-	const groupStories = storiesOf(demoGroup.title, module).addDecorator(withKnobs);
+	const groupStories = storiesOf(demoGroup.title, module).addDecorator(
+		withKnobs({ escapeHTML: false })
+	);
 
 	// Loop through the demos for the group
 	demoGroup.demos.forEach(demo => {