+++ packages/react/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.30.6](https://github.com/IBM/carbon-charts/compare/v0.30.5...v0.30.6) (2020-03-30)
+
+**Note:** Version bump only for package @carbon/charts-react
+
+
+
+
+
+## [0.30.5](https://github.com/IBM/carbon-charts/compare/v0.30.4...v0.30.5) (2020-03-30)
+
+**Note:** Version bump only for package @carbon/charts-react
+
+
+
+
+
+## [0.30.4](https://github.com/IBM/carbon-charts/compare/v0.30.3...v0.30.4) (2020-03-27)
+
+**Note:** Version bump only for package @carbon/charts-react
+
+
+
+
+
 ## [0.30.3](https://github.com/IBM/carbon-charts/compare/v0.30.2...v0.30.3) (2020-03-24)
 
 **Note:** Version bump only for package @carbon/charts-react